requirements.txt
browser-use>=0.1.0,<1.0.0
langchain-community>=0.0.1,<1.0.0
aiohttp>=3.9.0,<4.0.0
aiofiles>=23.0.0,<24.0.0
python-dotenv>=1.0.0,<2.0.0
